Revert "refactor: use typescript"
This reverts commit 5dd177fac73b87f691f5c97d5e616f32b4e65bc0.
