+++ package.json
@@ -8,30 +8,25 @@
   "license": "MIT",
   "private": false,
   "scripts": {
-    "dev": "next dev",
+    "dev": "next",
     "build": "next build",
     "start": "next start"
   },
   "dependencies": {
-    "@types/react-custom-scrollbars": "^4.0.6",
     "ajv": "6.5.5",
     "axios": "^0.18.1",
     "cookie": "^0.3.1",
     "date-fns": "^1.29.0",
     "immer": "^1.8.0",
-    "next": "^9.1.4",
-    "react": "^16.12.0",
+    "next": "^7.0.2",
+    "react": "^16.6.3",
     "react-custom-scrollbars": "^4.2.1",
-    "react-dom": "^16.12.0",
-    "use-immer": "^0.3.5"
+    "react-dom": "^16.6.3"
   },
   "devDependencies": {
-    "@types/node": "^12.12.12",
-    "@types/react": "^16.9.12",
     "husky": "^1.2.0",
     "prettier": "^1.15.2",
-    "pretty-quick": "^1.8.0",
-    "typescript": "^3.7.2"
+    "pretty-quick": "^1.8.0"
   },
   "resolutions": {
     "ajv": "6.5.5"
